refactor: rename builder react and sdk client
