+++ src/components/app/builder.tsx
@@ -2,13 +2,13 @@
 
 import '@/utils/web/builder/registerComponents'
 
-import { ComponentProps } from 'react'
+import type { ComponentProps } from 'react'
 import { BuilderComponent, useIsPreviewing } from '@builder.io/react'
 import { notFound } from 'next/navigation'
 
-import { maybeInitCMSClient } from '@/utils/web/builder/clientCMS'
+import { maybeInitBuilderReactClient } from '@/utils/web/builder'
 
-maybeInitCMSClient()
+maybeInitBuilderReactClient()
 
 type BuilderPageProps = ComponentProps<typeof BuilderComponent> & {
   type: 'section' | 'page'
