+++ src/utils/server/builder/models/data/constants.ts
@@ -1,0 +1,4 @@
+export enum DataModelIdentifiers {
+  EVENTS = 'events',
+  QUESTIONNAIRE = 'questionnaire',
+}
