+++ src/utils/server/builder/models/data/events.ts
@@ -1,7 +1,8 @@
 import * as Sentry from '@sentry/nextjs'
 import pRetry from 'p-retry'
 
-import { serverCMS } from '@/utils/server/builder/serverCMS'
+import { builderSDKClient } from '@/utils/server/builder'
+import { DataModelIdentifiers } from '@/utils/server/builder/models/data/constants'
 import { SWCEvents, zodEventSchemaValidation } from '@/utils/shared/getSWCEvents'
 import { getLogger } from '@/utils/shared/logger'
 import { NEXT_PUBLIC_ENVIRONMENT } from '@/utils/shared/sharedEnv'
@@ -12,7 +13,7 @@ export async function getEvent(eventSlug: string, state: string) {
   try {
     const entry = await pRetry(
       () =>
-        serverCMS.get('events', {
+        builderSDKClient.get(DataModelIdentifiers.EVENTS, {
           query: {
             data: {
               slug: eventSlug,
@@ -52,7 +53,7 @@ const LIMIT = 100
 async function getAllEventsWithOffset(offset: number) {
   return await pRetry(
     () =>
-      serverCMS.getAll('events', {
+      builderSDKClient.getAll(DataModelIdentifiers.EVENTS, {
         query: {
           ...(NEXT_PUBLIC_ENVIRONMENT === 'production' && { published: 'published' }),
           data: {
