+++ src/utils/server/builder/models/data/questionnaire.ts
@@ -1,7 +1,8 @@
 import * as Sentry from '@sentry/nextjs'
 import pRetry from 'p-retry'
 
-import { serverCMS } from '@/utils/server/builder/serverCMS'
+import { builderSDKClient } from '@/utils/server/builder'
+import { DataModelIdentifiers } from '@/utils/server/builder/models/data/constants'
 import { zodQuestionnaireSchemaValidation } from '@/utils/shared/getSWCQuestionnaire'
 import { getLogger } from '@/utils/shared/logger'
 
@@ -10,7 +11,7 @@ export async function getQuestionnaire(DTSISlug: string) {
   try {
     const entry = await pRetry(
       () =>
-        serverCMS.get('questionnaire', {
+        builderSDKClient.get(DataModelIdentifiers.QUESTIONNAIRE, {
           query: {
             data: {
               dtsiSlug: DTSISlug,
