+++ src/utils/web/builder/index.ts
@@ -1,0 +1,2 @@
+export { maybeInitBuilderReactClient } from './maybeInitBuilderReactClient'
+export type { BuilderComponentBaseProps } from './types'
